requirements.txt
fastapi
uvicorn
pillow
tensorflow
numpy
python-dotenv
langchain
langchain-google-genai
huggingface_hub
httpx
python-multipart